+++ examples/complete/main.tf
@@ -11,8 +11,7 @@ module "ssh_key_generator" {
   file_permission      = "600"
 }
 
-
-## Update complete example with new features
+##TODO: Update complete example with new features
 module "complete" {
   source = "../../"
 
@@ -24,7 +23,7 @@ module "complete" {
   include_preview     = true
   tags                = local.tags
 
-  ## Identity
+  ##? Identity
   identity_type                              = "UserAssigned"
   identity_ids                               = [azurerm_user_assigned_identity.aks.id]
   kubelet_identity_enabled                   = true
@@ -32,7 +31,7 @@ module "complete" {
   kubelet_identity_client_id                 = azurerm_user_assigned_identity.kubelet.client_id
   kubelet_identity_object_id                 = azurerm_user_assigned_identity.kubelet.principal_id
 
-  ## Default node pool
+  ##? Default node pool
   default_node_pool_name                = "system"
   default_node_pool_enable_auto_scaling = true
   default_node_pool_vm_size             = "standard_d2ds_v5"
@@ -40,7 +39,7 @@ module "complete" {
   default_node_pool_max_count           = 2
   default_node_pool_max_pods            = 110
 
-  ## additional_node_pools
+  ##? additional_node_pools
   additional_node_pools = {
     "nodepool01" = {
       vm_size             = "standard_d2ds_v5"
@@ -51,12 +50,12 @@ module "complete" {
     }
   }
 
-  ## Api service access profile
+  ##? Api service access profile
   enable_api_server_access_profile    = true
   vnet_integration_enabled            = true
   api_server_access_profile_subnet_id = azurerm_subnet.aks_api.id
 
-  ## Networking
+  ##? Networking
   vnet_subnet_id      = azurerm_subnet.aks_node.id
   network_plugin      = "azure"
   network_plugin_mode = "overlay"
@@ -66,24 +65,24 @@ module "complete" {
   network_policy      = "calico"
   # ebpf_data_plane     = "cilium"
 
-  ## Monitor Diagnostics Settings
+  ##? Monitor Diagnostics Settings
   enable_cluster_log_monitor_diagnostic = true
   monitor_diagnostic_storage_account_id = azurerm_storage_account.aks.id
   enable_cluster_all_category_group_log = true
   enable_cluster_all_metrics            = true
 
-  ## Azure Active Directory
+  ##? Azure Active Directory
   local_account_disabled           = true
   aad_rbac_enabled                 = true ## Enable the feature for Azure RBAC with AKS
   aad_rbac_managed                 = true ## Manged RBAC
   aad_azure_rbac_enabled           = true ## Azure AAD and Azure RBAC ( No K8s RBAC )
   aad_rbac_managed_admin_group_ids = []   ## require permissions to create groups , empty good for testing
 
-  ## Workload Identity
+  ##? Workload Identity
   workload_identity_enabled = true
   oidc_issuer_enabled       = true
 
-  ## Flux
+  ##? Flux
   enable_fluxcd_az_providers                     = true ## Enable the providers configurations required for flux extension.
   enable_fluxcd                                  = true
   fluxcd_extension_name                          = "infrastructure"
@@ -106,9 +105,16 @@ module "complete" {
       sync_interval_in_seconds = 60
     }
   ]
-  ## Extras
+
+  ##? Extras
   image_cleaner_enabled                               = true
   run_command_enabled                                 = true
   key_vault_secrets_provider_secret_rotation_enabled  = true
   key_vault_secrets_provider_secret_rotation_interval = true
+
+  ##? Explicit Dependency
+  depends_on = [
+    azurerm_role_assignment.kubelet_operator,
+    azurerm_role_assignment.aks_mi_network_contributor
+  ]
 }